refactor: bootstrap the tests for the incident model
