+++ internal/domain/monitor/incident_test.go
@@ -1,0 +1,7 @@
+package monitor_test
+
+import "testing"
+
+func TestName(t *testing.T) {
+
+}
