+++ internal/domain/monitor/repository.go
@@ -14,3 +14,10 @@ type Repository interface {
 	UpdateForCheck(ctx context.Context, fn func(foundMonitors []*Monitor) error) error
 	SaveCheckResult(ctx context.Context, ID domain.ID, checkResult *CheckResult) error
 }
+
+type IncidentRepository interface {
+	FindByID(ctx context.Context, id domain.ID) (*Incident, error)
+	Create(ctx context.Context, incident *Incident) error
+	Update(ctx context.Context, id domain.ID, fn func(incident *Incident) error) error
+	AppendIncidentAction(ctx context.Context, incidentID domain.ID, incidentAction *IncidentAction) error
+}
